slightly restructure the project and clean up some of the files
