+++ QuizWorld.Infrastructure/Filters/GuestsOnly/GuestsOnlyFilter.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
-using QuizWorld.Infrastructure.Data.Contracts;
+using QuizWorld.Web.Contracts.JsonWebToken;
 using System;
 using System.Collections.Generic;
 using System.Linq;
