+++ QuizWorld.Infrastructure/QuizWorld.Infrastructure.csproj
@@ -30,6 +30,7 @@
   <ItemGroup>
     <ProjectReference Include="..\QuizWorld.Common\QuizWorld.Common.csproj" />
     <ProjectReference Include="..\QuizWorld.ViewModels\QuizWorld.ViewModels.csproj" />
+    <ProjectReference Include="..\QuizWorld.Web.Contracts\QuizWorld.Web.Contracts.csproj" />
   </ItemGroup>
 
 </Project>
