+++ QuizWorld.Web/QuizWorld.Web.csproj
@@ -28,6 +28,8 @@
     <ProjectReference Include="..\QuizWorld.Common\QuizWorld.Common.csproj" />
     <ProjectReference Include="..\QuizWorld.Infrastructure\QuizWorld.Infrastructure.csproj" />
     <ProjectReference Include="..\QuizWorld.ViewModels\QuizWorld.ViewModels.csproj" />
+    <ProjectReference Include="..\QuizWorld.Web.Contracts\QuizWorld.Web.Contracts.csproj" />
+    <ProjectReference Include="..\QuizWorld.Web.Services\QuizWorld.Web.Services.csproj" />
   </ItemGroup>
 
 </Project>
